fix: route Gemini key fetch failures to Crashlytics and fix alpha/beta versionCode conflict
- SyncedAiAssistantSettings: inject AppLogger and record non-fatal via Crashlytics when
  all remote key fetch retries are exhausted (previously Kermit-only, invisible on dashboard).
  GeminiKeyRpcDecoder decode failures (get_gemini_api_key_decode_failed) are now surfaced
  as non-fatal events with the tag 'SyncedAiSettings'.

- androidApp/build.gradle.kts: separate debug/alpha versionCode formula.
  Debug builds now use 1_000_000 + run_number (≥ 1,000,631) instead of the
  100_000 + run_number*3 + 0 formula shared with release tracks (≤ ~120,000).
  A tester with the Play Store beta installed can now install the Firebase alpha debug
  APK without Android blocking it as a versionCode downgrade.

- Compiler warnings eliminated: quadraticBezierTo→quadraticTo in GroceryItemRow,
  redundant toInt() in WeekCalendar, @file:OptIn(ExperimentalCoroutinesApi) on
  HomeScreenModel and NutritionScreenModel, unnecessary safe-call and Elvis in
  NutritionAiAssistantContent, statusBarColor/navigationBarColor deprecated API guarded
  behind Build.VERSION_CODES.VANILLA_ICE_CREAM check in SystemBars.android.kt.

- New test: SyncedAiAssistantSettingsTest.refreshFromRemote_recordsNonFatalViaCrashlytics_whenAllFetchAttemptsFail

+++ composeApp/src/commonMain/kotlin/app/mymultiverse/ammo/domain/nutrition/WeekCalendar.kt
@@ -66,7 +66,7 @@ object WeekCalendar {
     ): Int? {
         if (weekKeyFor(today) != weekKey) return null
         val start = LocalDate.parse(weekKey)
-        val index = (today.toEpochDays() - start.toEpochDays()).toInt()
+        val index = today.toEpochDays() - start.toEpochDays()
         return index.takeIf { it in 0 until WeeklyMealPlan.DAYS_IN_WEEK }
     }
